+++ src/standalone/Soil/energy_hydrology.jl
@@ -198,8 +198,8 @@ function ClimaLand.make_compute_exp_tendency(
     function compute_exp_tendency!(dY, Y, p, t)
         z = model.domain.fields.z
 
-        p.soil.∫S_θ_liq_dz .= 0
-        p.soil.∫S_ρe_int_dz .= 0
+        dY.soil.∫F_vol_liq_water_dt .= 0
+        dY.soil.∫F_e_dt .= 0
 
         # Don't update the prognostic variables we're stepping implicitly
         dY.soil.ϑ_l .= 0
@@ -218,13 +218,13 @@ function ClimaLand.make_compute_exp_tendency(
             p,
             z,
         )
-        # Source terms, also update the vertical integral of the source: p.soil.∫S_θ_liq_dz, p.soil.∫S_ρe_int_dz
-        # These change ∫S and dY by +=, which is why we zero them out above.
+        # Explicitly treated source terms
+        # These change dY by +=, which is why we ".=" them above
         for src in model.sources
-            ClimaLand.source!(dY, src, Y, p, model)
+            if src.explicit
+                ClimaLand.source!(dY, src, Y, p, model)
+            end
         end
-        dY.soil.∫F_vol_liq_water_dt .= p.soil.∫S_θ_liq_dz # source terms are stepped explicitly
-        dY.soil.∫F_e_dt .= p.soil.∫S_ρe_int_dz # source terms are stepped explicitly
     end
     return compute_exp_tendency!
 end
@@ -291,8 +291,15 @@ function ClimaLand.make_compute_imp_tendency(
                 ) * gradc2f(p.soil.ψ + z),
             )
 
-        # Don't update the prognostic variables we're stepping explicitly
         @. dY.soil.θ_i = 0
+
+        # Source terms
+        # These change dY by +=, which is why we ".=" above
+        for src in model.sources
+            if !src.explicit
+                ClimaLand.source!(dY, src, Y, p, model)
+            end
+        end
     end
     return compute_imp_tendency!
 end
@@ -493,8 +500,6 @@ of `EnergyHydrology`.
 ClimaLand.auxiliary_vars(soil::EnergyHydrology) = (
     :total_water,
     :total_energy,
-    :∫S_θ_liq_dz,
-    :∫S_ρe_int_dz,
     :K,
     :ψ,
     :θ_l,
@@ -521,8 +526,6 @@ ClimaLand.auxiliary_types(soil::EnergyHydrology{FT}) where {FT} = (
     FT,
     FT,
     FT,
-    FT,
-    FT,
     boundary_var_types(
         soil,
         soil.boundary_conditions.top,
@@ -536,8 +539,6 @@ ClimaLand.auxiliary_types(soil::EnergyHydrology{FT}) where {FT} = (
 )
 
 ClimaLand.auxiliary_domain_names(soil::EnergyHydrology) = (
-    :surface,
-    :surface,
     :surface,
     :surface,
     :subsurface,
@@ -647,9 +648,12 @@ end
 """
     PhaseChange{FT} <: AbstractSoilSource{FT}
 
-PhaseChange source type.
+PhaseChange source type; treated explicitly in all
+prognostic variables.
 """
-struct PhaseChange{FT} <: AbstractSoilSource{FT} end
+@kwdef struct PhaseChange{FT} <: AbstractSoilSource{FT}
+    explicit::Bool = true
+end
 
 """
      source!(dY::ClimaCore.Fields.FieldVector,
@@ -659,7 +663,8 @@ struct PhaseChange{FT} <: AbstractSoilSource{FT} end
              model
              )
 
-Computes the source terms for phase change.
+Computes the source terms for phase change
+explicitly in time.
 
 """
 function ClimaLand.source!(
@@ -728,7 +733,7 @@ function ClimaLand.source!(
             _grav,
         )
 
-    # These source/sink terms conserve total water, so we do not include them in ∫S_θ_liq_dz because their
+    # These source/sink terms conserve total water, so we do not include them when checking conservation because their
     # contribution is zero by design.
 end
 
@@ -737,10 +742,12 @@ end
     SoilSublimation{FT} <: AbstractSoilSource{FT}
 
 Soil Sublimation source type. Used to defined a method
-of `ClimaLand.source!` for soil sublimation.
+of `ClimaLand.source!` for soil sublimation; treated implicitly
+in ϑ_l, ρe_int but explicitly in θ_i.
 """
-struct SoilSublimation{FT} <: AbstractSoilSource{FT} end
-
+@kwdef struct SoilSublimation{FT} <: AbstractSoilSource{FT}
+    explicit::Bool = false
+end
 """
      source!(dY::ClimaCore.Fields.FieldVector,
              src::SoilSublimation{FT},
@@ -767,7 +774,7 @@ function ClimaLand.source!(
     @. dY.soil.θ_i +=
         -p.soil.turbulent_fluxes.vapor_flux_ice * _ρ_l / _ρ_i *
         heaviside(z + 2 * Δz_top) / (2 * Δz_top) # only apply to top layer, recall that z is negative
-    @. p.soil.∫S_θ_liq_dz += -p.soil.turbulent_fluxes.vapor_flux_ice # The integral of the source is designed to be this
+    @. dY.soil.∫F_vol_liq_water_dt += -p.soil.turbulent_fluxes.vapor_flux_ice # The integral of the source is designed to be this
 end
 
 ## The functions below are required to be defined